requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.26.0
fastapi>=0.109.0
uvicorn>=0.27.0
sentence-transformers>=2.7.0
streamlit>=1.29.0

